Shader review/update + resource layout patches
- Adjusted PT cache coherency + memory usage
- Some shader syntax updates for SM6
- Implemented resource contexts to help match cpu descriptor layouts to layouts expected by shaders, and standardize layouts generally to simplify PSO generation
- Flattened + standardized PT descriptor layouts to match the single standard layout given by the cpu-side [RENDER] resource context

+++ Athru/Athru Utilities/UtilityLib/UtilityLib/AppGlobals.h
@@ -23,7 +23,7 @@ namespace TimeStuff
 
 	inline float FPS()
 	{
-		return std::round(1.0f / deltaTime());
+		return (float)round(1.0f / deltaTime());
 	}
 
 	inline float time()
